Extended Keycodes
Change keycodes from byte to 2 bytes to adopt consumer/system pages

+++ common/keymap.h
@@ -23,12 +23,12 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 
 /* keycode in specific layer */
-uint8_t keymap_get_keycode(uint8_t layer, uint8_t row, uint8_t col);
+uint16_t keymap_get_keycode(uint8_t layer, uint8_t row, uint8_t col);
 
 /* layer to move during press Fn key */
 uint8_t keymap_fn_layer(uint8_t fn_bits);
 
 /* keycode to send when release Fn key without using */
-uint8_t keymap_fn_keycode(uint8_t fn_bits);
+uint16_t keymap_fn_keycode(uint8_t fn_bits);
 
 #endif
