+++ common/layer.h
@@ -24,7 +24,7 @@ extern uint8_t default_layer;
 extern uint8_t current_layer;
 
 /* return keycode for switch */
-uint8_t layer_get_keycode(uint8_t row, uint8_t col);
+uint16_t layer_get_keycode(uint8_t row, uint8_t col);
 
 /* process layer switching */
 void layer_switching(uint8_t fn_bits);
